Relocate navigationMenu and refactor header and navigationMenu tests
